Fully featured sequences load just like the single exr images now.  Separated sequence loader into its own node due to js complexity.  Adjusted all layers to use comfy caching logic method
